+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - After editing a match, the view will now scroll to that match on the Matches page
 - Added link on the About page to let you check for a newer version of the app [#18](https://github.com/cheshire137/competiwatch-desktop/issues/18)
 - Added "Export matches" link to the Matches page to allow downloading your match history for that account and season [#32](https://github.com/cheshire137/competiwatch-desktop/issues/32)
+- Added "Export Matches" menu item to the Tools menu when viewing a particular account's match history
 
 ## Version 0.1.6